# openapi/api.json
{
  "openapi": "3.0.0",
  "paths": {
    "/api/v1/users": {
      "post": {
        "operationId": "registerUser",
        "summary": "新規登録",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "users"
        ]
      }
    },
    "/api/v1/authentication/login": {
      "post": {
        "operationId": "login",
        "summary": "ログイン",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "authentication"
        ]
      }
    },
    "/api/v1/authentication/logout": {
      "post": {
        "operationId": "logout",
        "summary": "ログアウト",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "authentication"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/authentication/me": {
      "get": {
        "operationId": "getMe",
        "summary": "ログイン済のユーザー情報を取得",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "authentication"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/comparisons": {
      "post": {
        "operationId": "createComparison",
        "summary": "比較情報を登録",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateComparisonDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateComparisonResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "comparisons"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "get": {
        "operationId": "getComparisons",
        "summary": "比較情報一覧を取得",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ComparisonResponseDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "comparisons"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/comparisons/{id}": {
      "get": {
        "operationId": "findComparison",
        "summary": "比較情報を取得",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ComparisonResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "comparisons"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "deleteComparison",
        "summary": "比較情報を削除",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "comparisons"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/comparisons/{id}/publish": {
      "put": {
        "operationId": "publishComparison",
        "summary": "比較情報を公開状態にする",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "comparisons"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/youtube/oauth/authorize": {
      "get": {
        "operationId": "getAuthorizeUrl",
        "summary": "認可画面のURLを取得",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthorizeUrlResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "youtube"
        ]
      }
    },
    "/api/v1/youtube/oauth": {
      "post": {
        "operationId": "exchangeToken",
        "summary": "アクセストークンを取得",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/YoutubeOauthDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExchangeTokenResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "youtube"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/youtube/videos": {
      "get": {
        "operationId": "fetchVideos",
        "summary": "本人がアップロードした動画一覧を取得",
        "parameters": [
          {
            "name": "x-youtube-access-token",
            "required": true,
            "in": "header",
            "description": "YouTubeアクセストークン",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/YoutubeVideoDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "youtube"
        ]
      }
    },
    "/api/v1/operation-log": {
      "put": {
        "operationId": "updateOperationLog",
        "summary": "操作ログを送信",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OperationLogDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "operationLog"
        ]
      }
    },
    "/api/v1/status": {
      "get": {
        "operationId": "getStatus",
        "summary": "未ログインユーザーも含めたユーザーの状態を取得",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatusResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "status"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    }
  },
  "info": {
    "title": "RunCheck API",
    "description": "RunCheck Backend API",
    "version": "1.0.0",
    "contact": {}
  },
  "tags": [],
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "CreateUserDto": {
        "type": "object",
        "properties": {
          "handleName": {
            "type": "string",
            "description": "ハンドルネーム"
          },
          "carType": {
            "type": "string",
            "description": "車種"
          },
          "email": {
            "type": "string",
            "description": "メールアドレス"
          },
          "password": {
            "type": "string",
            "description": "パスワード"
          }
        },
        "required": [
          "handleName",
          "carType",
          "email",
          "password"
        ]
      },
      "UserResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "ユーザーID"
          },
          "name": {
            "type": "string",
            "description": "ネーム"
          }
        },
        "required": [
          "id",
          "name"
        ]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "メールアドレス"
          },
          "password": {
            "type": "string",
            "description": "パスワード"
          },
          "remember": {
            "type": "boolean",
            "description": "ログイン状態を保持"
          }
        },
        "required": [
          "email",
          "password",
          "remember"
        ]
      },
      "LoginResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "ユーザーID"
          },
          "name": {
            "type": "string",
            "description": "ネーム"
          },
          "accessToken": {
            "type": "string",
            "description": "アクセストークン"
          }
        },
        "required": [
          "id",
          "name",
          "accessToken"
        ]
      },
      "CreateComparisonDto": {
        "type": "object",
        "properties": {
          "category": {
            "type": "string",
            "description": "カテゴリ"
          },
          "memo": {
            "type": "string",
            "description": "メモ"
          },
          "title": {
            "type": "string",
            "description": "タイトル"
          },
          "video1Url": {
            "type": "string",
            "description": "動画1 URL"
          },
          "video1TimeSt": {
            "type": "number",
            "description": "動画1 開始時間"
          },
          "video1VideoType": {
            "type": "string",
            "description": "動画1 動画タイプ",
            "enum": [
              "1",
              "2"
            ]
          },
          "video2Url": {
            "type": "string",
            "description": "動画2 URL"
          },
          "video2TimeSt": {
            "type": "number",
            "description": "動画2 開始時間"
          },
          "video2VideoType": {
            "type": "string",
            "description": "動画2 動画タイプ",
            "enum": [
              "1",
              "2"
            ]
          },
          "anonymous": {
            "type": "boolean",
            "description": "匿名投稿"
          }
        },
        "required": [
          "video1Url",
          "video1TimeSt",
          "video1VideoType",
          "video2Url",
          "video2TimeSt",
          "video2VideoType",
          "anonymous"
        ]
      },
      "CreateComparisonResponseDto": {
        "type": "object",
        "properties": {
          "comparisonId": {
            "type": "string",
            "description": "比較ID"
          }
        },
        "required": [
          "comparisonId"
        ]
      },
      "ComparisonResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "比較ID"
          },
          "category": {
            "type": "string",
            "description": "カテゴリ"
          },
          "memo": {
            "type": "string",
            "description": "メモ"
          },
          "title": {
            "type": "string",
            "description": "タイトル"
          },
          "video1Url": {
            "type": "string",
            "description": "動画1 URL"
          },
          "video1TimeSt": {
            "type": "number",
            "description": "動画1 開始時間"
          },
          "video1VideoType": {
            "type": "string",
            "description": "動画1 動画タイプ",
            "enum": [
              "1",
              "2"
            ]
          },
          "video2Url": {
            "type": "string",
            "description": "動画2 URL"
          },
          "video2TimeSt": {
            "type": "number",
            "description": "動画2 開始時間"
          },
          "video2VideoType": {
            "type": "string",
            "description": "動画2 動画タイプ",
            "enum": [
              "1",
              "2"
            ]
          },
          "anonymous": {
            "type": "boolean",
            "description": "匿名投稿"
          }
        },
        "required": [
          "id",
          "video1Url",
          "video1TimeSt",
          "video1VideoType",
          "video2Url",
          "video2TimeSt",
          "video2VideoType",
          "anonymous"
        ]
      },
      "AuthorizeUrlResponseDto": {
        "type": "object",
        "properties": {
          "redirectUrl": {
            "type": "string",
            "description": "認可画面のリダイレクト用URL"
          }
        },
        "required": [
          "redirectUrl"
        ]
      },
      "YoutubeOauthDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "description": "OAuth認可コード"
          }
        },
        "required": [
          "code"
        ]
      },
      "ExchangeTokenResponseDto": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string",
            "description": "YouTubeアクセストークン"
          }
        },
        "required": [
          "accessToken"
        ]
      },
      "YoutubeVideoDto": {
        "type": "object",
        "properties": {
          "url": {
            "type": "string",
            "description": "URL"
          },
          "title": {
            "type": "string",
            "description": "タイトル"
          },
          "description": {
            "type": "string",
            "description": "説明"
          },
          "thumbnailUrl": {
            "type": "string",
            "description": "サムネ画像のURL"
          }
        },
        "required": [
          "url",
          "title",
          "description",
          "thumbnailUrl"
        ]
      },
      "OperationLogDto": {
        "type": "object",
        "properties": {
          "operationCd": {
            "type": "string",
            "description": "操作コード",
            "enum": [
              "1",
              "2",
              "3",
              "4",
              "5",
              "6",
              "7",
              "8",
              "9",
              "10",
              "11",
              "12",
              "13",
              "14",
              "15",
              "16",
              "17",
              "18",
              "19",
              "20",
              "21",
              "22",
              "23",
              "24",
              "25",
              "26",
              "27",
              "28",
              "29",
              "30"
            ]
          }
        },
        "required": [
          "operationCd"
        ]
      },
      "StatusUserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "ユーザーID"
          },
          "name": {
            "type": "string",
            "description": "ネーム"
          }
        },
        "required": [
          "id",
          "name"
        ]
      },
      "StatusResponseDto": {
        "type": "object",
        "properties": {
          "isLogined": {
            "type": "boolean",
            "description": "ログイン済みかどうか"
          },
          "isYoutubeAuthroized": {
            "type": "boolean",
            "description": "YouTube認証済みかどうか"
          },
          "user": {
            "description": "ユーザー情報",
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/StatusUserDto"
              }
            ]
          }
        },
        "required": [
          "isLogined",
          "isYoutubeAuthroized"
        ]
      }
    }
  }
}